removed dependency
the mobius widget module in the mobius core is use in one file for an error response which so far isn't firing at this point so commented it out.
This likely means the widget modules will need improving and consolidating correctly. for now i have commented out the section and marked it as TODO