found last circular dependency
We have a few pitfalls to solve between modules and written out a couple of notes on addressing the issues at hand.

As the solutions note, it will require new components to resolve this, so as we now have users downloading and forking the current project, i merge the branches and work on addressing this.