Creating a simple agent view variation
While the performance for animated agents is good, lower hardware specifications may struggle with it.

I have added a copy of the current system and turned off the animations and dynamic textures for testing and have tested this by switching the processor import lines for this asset and validated that updates occur successfully.

TODO:
* need to implement a toggle or a auto toggle when FPS drops below a threshold.
* need to implement code logic to turn off expensive animation lookup setting changes
* implement new variables in existing system to only turn off components and swap meshes
* Current mesh uses 3 materials and should only use 1 and be a material instance
* need to scale the mesh for children, no gender currently exists for this and likely wont