simplified mass observer processors and corrected logic
Previously the observer processors were ignoring execution logic due to how the processors were firing on the observed type add operation. To correct this I have reimplemented tags so that render data will only be processed when the pedestrian observer gives the render tag.

by stripping movement logic from the render MOP, we are no longer performing double assignments to the entities and keeping each processor to the specific task it is assigned