Execution logic now matches new fragment arrangement
Have organized the entity info fragment into 3 separate types, one that stores info, movement and rendering data. The processors and subsystems now embody this new logic as well.

TODO:
Update comments in relative areas so methods provide the correct information and the execution logic is documented correctly

Investigate:
the entity ID may not be required as the loop logic in most cases no longer requires it? There is also areas of processors that could be reduced or refactored to increase readability and performance