Initialize data bug
Because the way entities are destroyed, the indexing assigned to new entities is offset by the previous amount, so when we changed from large to small it would cause indexing issues and misalignment problems. this has since been resolved and will undergo further testing tomorrow to ensure it is fixed.