Add AgentInfoDisplay widget and mesh rendering support WIP
Introduces the UAgentInfoDisplay widget and supporting SPedestrianAgentMeshWidget for in-world agent info visualization. Adds new assets for AgentInfoDisplay, updates build dependencies to include Niagara, and configures class redirects. Also includes minor code comments and updates to related UI and level assets.

TODO: There is an offset issue that is causing the widget to be misaligned