Add MobiusControllerSubsystem and UI stats display for agents
Introduces UMobiusControllerSubsystem to manage player controller and mouse-based world interaction, including line tracing from mouse position. Adds FDisplayEntityDetailsTag and showPedestrianStats to support displaying agent UI stats. Implements DisplayAgentUIStatsProcessor for updating UI with selected agent data. Updates PedestrianCollisionProcessor to use MobiusControllerSubsystem for mouse interaction and entity selection. Refactors and cleans up related MassAI processors and fragments. Updates config for class redirects and moves MobiusController.h to public.

The current implementation works as a rough prototype for the hover logic, the next step is to create the click lock widget logic so we always display the information on an agent to the user.

After all prototype logic is completed we need to perform a code cleanup refactor on these new classes  and fragments but also the biggest task is optimization, while having all the collisions moving will be desirable we need to turn features off as moving large amounts of capsules is highly inefficient and causing significant drop in performance. Our 5000 agent test sample run 25fps to 35fps now whereas before it was 115 to 125fps