Refactor agent UI display and selection logic
Introduces new SAgentFollowIndicator and refactors PedestrianAgentMeshWidget to SPedestrianAgentHoverMeshWidget for improved agent hover and selection visualization. Updates StatisticSubsystem to track selected agent data and adds related delegates. PedestrianDataDisplay replaces SelectedAgentDisplay, now updating via subsystem delegate. FieldAndTextWidget and SFieldAndTitleText now support separate title and field text styles. AgentInfoDisplay now manages hover and selected agent widgets and mesh IDs. MassAI processors updated to distinguish between hovered and selected agents for UI display. Adds new style assets for pedestrian data display fields and titles. The follow widget now has a small bounce to so it is easier to identify to users

WIP Update:
The core functionality works for this feature, the remaining tasks predominantly focus on designing the behaviour on when to show the follow agent data on screen, but also a method to activate the hover while ctrl key is held.
The biggest tasks left for this feature is to refactor the code and refine it.