Refactor flow counter for thread safety and add analytics
Replaced TAtomic with std::atomic for flow counter and Z bounds to improve thread safety (TAtomic is deprecated in Unreal and need to use std::atomic). Introduced a thread-safe queue for agent bucket assignment and deferred bucket mutation to the game thread. Added multiple static flow analytics methods to UStatisticSubsystem for density, flow, evacuation time, and Weidmann speed calculations. Updated ButtonWithText to support dynamic text changes and improved its widget construction.

TODO:
need to optimize UI for flow counters and the text and widget size needs to be dynamic as it can dynamically change the number of containers which adds or removes text, also the widget component scales up as we get further away the widget sizing needs to be capped as it shouldn't become so large it covers the simulation or other widgets
Because the flow counter widgets are beginning to scale up with more functionality and potentially complex equations, it needs to be moved into from a blueprint class to a CPP class, which can also aid the communication task and reduce references from one class to the other.

Lastly the equations for flow rate etc, need to be verified that they are correct as they are ChatGPT generated and may be incorrect or redundant