Update .gitignore for macOS and add Assimp plugin info
Expanded .gitignore to exclude common macOS system and Xcode files. Updated README with instructions for building the UE4_Assimp plugin, including relevant links and usage notes.

TODO: the readme file needs to be improved on instructions on the assimp plugin and update old details when assimp was manually in place to point towards the new assimp plugin. Also the instructions need to include mac building and troubleshooting issues that may occur, one issue identified is generating xcode solution files can be problematic and terminal prompts may be needed to generate the xcode for compiling